Remove Terraform file and update lint tasks
The terraform.tf file is removed from the project template. Additionally, changes are made to the linting tasks in lint.task.yml. A new lint all task is added that does not exit on failure, and the terraform fmt check in the lint all task script is now set to autofix formatting issues.